end message + attempt to expand foresight to depth 3
Pressing 'y' will execute the prototype of a more efficient algorithm. The current problem is that once the theoretical moves are conducted on the theory board (with no problem), evaluating the theory board gives the wrong result. 

I'll have go into debugging mode again. Perhaps there are some errors in the depth2 brain that need fixing.